Add word translation and TTS on hover
Use Intl.Segmenter for word segmentation. It's not availible in Firefox
as of now. But there is a working option of using an Segment-it
npm package. Should use it when making extension work in Firefox,

Use CC-CEDICT - CN-EN dictionary for word translations. Load it into
memory for fast translations.

For TTS use built-in SpeechSyntesis browser API.