Add translation speech synthesys
I was getting quite a bit of eye strain when using extension as is. The
reason is me constantly looking up at translation tooltip for each word.
Solution could be to text-to-speech the translation so there is no need
to move the eyes.

I shall try using after a little rest to see if it makes it better. Right
now I think the tooltip appearing is annoying too, so maybe it has to go.